docs: rename components/content/layout file
