+++ docs/app/docs/components/[slug]/content/items/index.ts
@@ -1,5 +1,5 @@
 import { concepts } from './concepts'
-import { layout } from './layout'
+import { layout } from './layout-items'
 import { components } from './components'
 import { utilities } from './utilities'
